Use WithTranslationAPIServiceTrait
