+++ layers/Engine/packages/component-model/src/Services/BasicServiceTrait.php
@@ -6,23 +6,11 @@ namespace PoP\ComponentModel\Services;
 
 use PoP\Hooks\Services\WithHooksAPIServiceTrait;
 use PoP\Root\Services\WithInstanceManagerServiceTrait;
-use PoP\Translation\TranslationAPIInterface;
-use Symfony\Contracts\Service\Attribute\Required;
+use PoP\Translation\Services\WithTranslationAPIServiceTrait;
 
 trait BasicServiceTrait
 {
     use WithInstanceManagerServiceTrait;
     use WithHooksAPIServiceTrait;
-
-    protected TranslationAPIInterface $translationAPI;
-
-    #[Required]
-    final public function setTranslationAPI(TranslationAPIInterface $translationAPI): void
-    {
-        $this->translationAPI = $translationAPI;
-    }
-    final protected function getTranslationAPI(): TranslationAPIInterface
-    {
-        return $this->translationAPI;
-    }
+    use WithTranslationAPIServiceTrait;
 }
