+++ layers/Engine/packages/translation/src/Services/WithTranslationAPIServiceTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoP\Translation\Services;
+
+use PoP\Translation\TranslationAPIInterface;
+use Symfony\Contracts\Service\Attribute\Required;
+
+trait WithTranslationAPIServiceTrait
+{
+    protected TranslationAPIInterface $translationAPI;
+
+    #[Required]
+    final public function setTranslationAPI(TranslationAPIInterface $translationAPI): void
+    {
+        $this->translationAPI = $translationAPI;
+    }
+    final protected function getTranslationAPI(): TranslationAPIInterface
+    {
+        return $this->translationAPI;
+    }
+}
